Added behaviour with sockets

+++ backend/initdb/initial.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-                                     id INTEGER PRIMARY KEY,
+                                     id BIGINT PRIMARY KEY,
                                      name TEXT DEFAULT '',
                                      balance BIGINT DEFAULT 0,
                                      wallet VARCHAR DEFAULT NULL
@@ -13,13 +13,13 @@ CREATE TABLE rooms (
 
 CREATE TABLE room_users (
                             room_id UUID REFERENCES rooms(id) ON DELETE CASCADE,
-                            user_id INTEGER,
+                            user_id BIGINT,
                             PRIMARY KEY (room_id, user_id)
 );
 
 CREATE TABLE transactions (
                               id SERIAL PRIMARY KEY,
-                              user_id INTEGER NOT NULL,
+                              user_id BIGINT NOT NULL,
                               amount BIGINT NOT NULL,
                               type smallint NOT NULL, -- 'withdrawal' / 'deposit'
                               tx_hash TEXT NOT NULL,
@@ -28,7 +28,7 @@ CREATE TABLE transactions (
 
 CREATE TABLE pending_deposits (
                                   id SERIAL PRIMARY KEY,
-                                  user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+                                  user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
                                   boc TEXT NOT NULL,
                                   amount BIGINT NOT NULL,
                                   retry_count INTEGER DEFAULT 0,
@@ -38,7 +38,7 @@ CREATE TABLE pending_deposits (
 
 CREATE TABLE game_history (
                               id SERIAL PRIMARY KEY,
-                              user_id INTEGER NOT NULL,
+                              user_id BIGINT NOT NULL,
                               bet BIGINT NOT NULL,
                               result bit NOT NULL,
                               created_at TIMESTAMP DEFAULT now()
